Add executor for parsing news
